Create advancedapproach.cs
To make the AI more efficient in calculating the imminent danger and shooting the nearest enemy, we can use an array of Enemy objects to store the enemy data (name and distance) and then sort it based on distance. By doing so, we can easily identify the nearest enemy and target it for shooting.
In this refactored version, I create an Enemy class to store the name and distance of each enemy. I use an array of Enemy objects to hold the enemy data. After reading the enemy data, I sort the array based on the distance of each enemy. The nearest enemy will always be the first element in the sorted array, so this shoot at that enemy.

By using this approach, the AI will react faster to any enemies, as it no longer needs to hard-code conditions to find the nearest enemy; instead, it automatically selects the nearest one based on the actual distances read from the input. This solution also allows for easy expansion if the number of enemies increases in the future.